requirements.txt
# Core Dependencies (already installed in Bosch environment)
streamlit==1.49.1
sqlalchemy==2.0.43
jinja2==3.1.6
python-docx==1.2.0
python-dotenv==1.1.1
pandas==2.3.2
fastapi==0.116.1
openai==1.106.1

# Additional packages for document management
openpyxl>=3.1.0  # Excel file handling
reportlab>=4.0.0  # PDF generation
PyPDF2>=2.0.0   # PDF manipulation
xlsxwriter>=3.0.0  # Advanced Excel features
pydantic>=2.0.0  # Data validation
uuid  # For unique IDs
datetime  # Date/time handling
pathlib  # File path handling

# Development dependencies
ipykernel
